+++ jaot_base/tests/__init__.py
@@ -4,3 +4,4 @@ from . import test_models
 from . import test_lifecycle
 from . import test_security
 from . import test_error_paths
+from . import test_adversarial